Move ClusterNameWidget
